rename to bunlde product
