+++ packages/api/resolvers/types/index.js
@@ -6,8 +6,8 @@ import Money from './money';
 import Media from './media';
 import SimpleProduct from './simple-product';
 import ConfigurableProduct from './configurable-product';
-import SetProduct from './set-product';
-import ProductSetItem from './product-set-item';
+import BundleProduct from './bundle-product';
+import ProductBundleItem from './product-bundle-item';
 import ProductMedia from './product-media';
 import ProductPrice from './product-price';
 import Dispatch from './dispatch';
@@ -86,10 +86,10 @@ export default {
   ProductVariationOption,
   ProductVariationAssignment,
   ProductVariationAssignmentVector,
-  ProductSetItem,
+  ProductBundleItem,
   SimpleProduct,
   ConfigurableProduct,
-  SetProduct,
+  BundleProduct,
   Assortment,
   Filter,
   FilterOption,
